Adding a test for nested works
Creates a work and adds N works as ordered members. This is to compare
performance against the collections test which uses reversed memberships
to add works to collections.

Note, the name of the test is "nested_collections" so that it aligns
with the choish test that does the same, only it is using collection
objects instead of works. The result, however, are the same since they
are simple Fedora resources.