Implement set of required work types, ref #348
Work types define how works are created, and should be chosen as a prior
step before creating the actual work.

Here we define a Valkyrie resource for the work type and allow it to be
assigned to the work as an id. We change the workflow for creating works
by explicitly linking to the existing work types available and render a
form with the work type already selected in a hidden field.

Note: this excludes the additional refactoring that assigns the fields
of the work type, and the form display that governs how the different
fields will be displayed for different work types.