Adds aria required attributes for the title input. Adds the required designation to the data dictionary ruby file and the csv.
Updates style sheet to add bottom margin to all inputs and make the submissions form look for a required designation and add CSS and required attributes to the field. Uses the label for the block, makes the lable titleized, and adds the critical required span. Updates some tests for the title field to be required and no longer just required to publish.

Switches required to publish to required in the fileds for change set so that the other change set tests pass without modification. We still want to test for the absence of title.