Refactors the default view so that the main h1 isn't hidden and isn't repeating the name of the application. Refactors the headings and layout for the search results page, so that we have a good document outline and that the content comes before the facets in the code.
Fixes insuficient color contrast for UI elements and overrides the layout helper for column sizes. Adds label for the file upload button and a table header for the data dictionary.

Replaces the search history table with an unorderd list. Consistent HTML for search history and saved searches. Rubocop fixes. Fixes the layout helper from blacklight so that the show page doesn't have the pull class to keep the tools column on the right.

Adds comments to the view files so that others may understand why we are overriding so many Blacklight views.