Delete one or more resources via confirmation process
Co-authored-by: Michael Tribone <mtribone@psu.edu>

Uses a new instance of Blacklight::Catalog as an interface for
searching, selecting, and then deleting one or more resources with a
confirmation step.

Delete requests from edit views are also routed to this interface where
the user can confirm the selection of resources to be deleted and
complete the process.