Import zip files as bags when uploaded via the GUI
When uploading a zip file to a new work in the GUI, the file should be
processed as a bag. This includes refactoring ImportFromZip so that it
will take either the zip name (as it does for csv import) or the full
path to the zip file (as it should for a GUI upload).

Co-authored-by Michael Tribone <mtribone@psu.edu>