Apply CSV metadata to FileSet
When importing works that contain file sets, read the metadata for each
file set from the provided csv. We can determine if a row of the csv is
a file set or work and then pass this information on to the transaction
which will apply the metadata when creating the file set.