Merge member of collection ids field into data dictionary
The member_of_collection_ids property was defined on works, but could
not be controlled using the data dictionary. This adds the field to the
data dictionary enabling the same UI elements such as "required." Now,
the field will be report as required the same way other fields that are
also defined as required.

Additional changes:

* defines a Valkyrie::ID field type
* assigns validations to change set from the data dictionary
* creates a collections controlled vocabulary
* member_of_collection_ids is no longer multiple
* Schema::WorkTypeConfiguration builds new fields using the change set

Co-authored-by: Michael Tribone <mtribone@psu.edu>