Use alternate_ids in place of identifier
Valkyrie has a feature where the attribute `alternate_ids` can function
like a different, but unique, Valkyrie id. This aligns well with CHO's
notion of a second unique identifier that is library-specific and not a
UUID like the default Valkyrie ID.

The consequence here is that the method must be named alternate_ids.
This may cause some confusion in the future since this will need to the
be column for the csv file when importing. We could examine using
aliases to avoid this.