Extract text when importing works from bags
Extracts text from preservation files that have text-based content and
creates an extracted text file in the file set. If the file set
already provides an extracted text file, or the preservation file does
not have text-based content, then no extraction is performed.

Corrects a bug where use type was not being correctly provided on
imported works. This enables extracted text content that is given
in the bag to be indexed on the file set.

New Feature: We can generate text content for files created via the bag
factory, or provide a specific file to be used in the bag. This is handy
if we want to provide a specific pdf file to use as the representative
for a work.