Display thumbnails in search results and show views
Uses customized presenters in Blacklight and additional changes to
views.

Adds Bootstrap spacing and sixing utilities to adjust the size of the thumbnail in relation to the content on various screen sizes.

Overrides _show_main_content and moves the pagination if search context from it to the show page. https://github.com/psu-libraries/cho/issues/124

Co-authored-by: Mike Tribone <mtribone@psu.edu>