Batch creation of Agent resources
Create and update agent resources via csv upload. Uses the existing
framework we have for works and extends to agents, using csv files to
create new agents or update existing ones. Additionally, it validates
agents so that no duplicate resources can be created that have the same
first and last name.

Since this was based largely on existing code for works, there was a
substantial amount of refactoring done to extract out common features.