Creates a new controller for the home page, adds an index partial, and updates the routes to have the root point to the new home index partial.
Moves the rendered search bar so that it may be positioned differently on the home page versus the secondary pages.

Addsp the secondary layout for the home page browse by and featured items, additional CSS for display, and a view test to check the presence of the secondary items. Fixed facet tests to start from search instead of home page

Added the background image to the home page, static starter images for the browse and featured sections, and adds Bootstrap columns for the section to make the breakpoints and images a little more predictable.

Adds the browse and featured images to the home page, swaps the inline block for Bootstrap columns for better image and column control at varying sizes, removes some styles for Bootstrap classes, removes the placeholder text.

Co-authored-by: Carolyn Cole <cam156@psu.edu>