Enabling private access to collections
Restricts collections to the users who have created them by adding a
"system creator" attribute to the collection. The attribute is not
present in the data dictionary but can be added to any collection or
resource.

The system creator is always a member of read users, to ensure they
have access to the resource. Additional changes were made to
enforce read access rights defined in a collection's access level.

Note: Edit rights are not controlled. This will be dealt with in a later
ticket.

Changes

* refactored transactions into AccessControls module
* updated transactions to make changes to access control attributes
* access control attributes added change sets
* default system user or current user is assigned to all new resources