Setting access rights in CSV
Adds support for setting access rights in a CSV when adding new
works to a collection.

Access rights is added to the data dictionary as a special kind of
property that is linked to a Repository::AccessLevel. Setting the kind
of access right will trigger behaviors such as limiting a resource
(collection or work) to registered or private users.

Changes:

* adding access_rights to data dictionary
* Add radio_button field type used only for access rights at this time
* Adding validation and controlled vocabularies
* Replace access_level with access_rights where appropriate
* Making config changes to dev and prod data dictionaries

Caveats:

The access_rights property is tightly coupled with the repository access
level concept, and functions will not work as expected unless this
coupling is maintained. For example, changing access_rights to a
different name would break these features. Future enhancements could
make this coupling more transparent or flexible.